Matrix Transfer Function ++++++++++++++++++++++++
Calculate CT_LTI_System.tf directly (instead of via poles and zeros). This
made a series of functions necessary that are able to handle matrices of
poly1d:

- minor(A, i, j)
- determinant(A)
- cofactorMat(A)
- polyDiag(polyList)

(Fixes #5)

SISO systems are not handled in a special manner anymore which lead to the
adaption of a number of unit tests.

Added second-order system to the library (Order2).