Fixed MIMO impulse- and step-response +++++++++++++++++++++++++++++++++++++
Fixed shapes of MIMO impulse- and step-responses to be
(lent, outputs, inputs). This is a bit different from the original
unittest which required (outputs, inputs, lent) but easier to plot:

plot(stepresponse.reshape(lent, -1))

Consequently, I have also transposed the shape of the frequency
response.

Further, I improved the docstring of tanuna.connect and changed the
argument naming to be consistent with CT_LTI_System.__connect__(...).